Update LandingTerrform.tf
Changed the S3 bucket permissions so only cloudfront can access the objects inside of it and not the other internet scoundrels. Added the pictures to this bucket, tomorrow I have to figure out how to get the pictures to load onto the page from portfolio.html file. 

I think I have to host a cloudfront distribution with the images that only this cloudfront distro can access. Maybe, maybe not. I'll figure it out tomorrow, right now I'm tired.